Move Wait functions into separate package
